requirements.txt
aiohttp==3.7.4.post0
apiai==1.2.3
arrow==1.2.2
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
beautifulsoup4==4.10.0
braceexpand==0.1.7
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
config==0.5.1
COVID19Py==0.3.0
Cybernator==0.6.3
DateTime==4.4
discord==1.7.3
discord-components==2.1.2
discord.py==1.7.3
discord-py-slash-command==3.0.3
disnake==2.4.0
ffmpeg==1.4
geojson==2.5.0
idna==3.3
import-expression==1.1.4
Jinja2==3.0.3
jinja2-time==0.2.0
jishaku==2.3.2
make==0.1.6.post2
MarkupSafe==2.1.0
multidict==6.0.2
nekos.py==1.0.3
numpy==1.22.3
pip==22.0.4
pymongo==4.0.2
pyowm==3.3.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
setuptools==58.1.0
six==1.16.0
soupsieve==2.3.1
typing_extensions==4.1.1
urllib3==1.26.8
wavelink==1.2.2
wheel==0.37.1
wikipedia==1.4.0
yarl==1.7.2
youtube-dl==2021.12.17
zope.interface==5.4.0